requirements.txt
streamlit
pandas
openpyxl
reportlab
google-generativeai